requirements.txt
absl-py==1.3.0
aiobotocore==2.4.2
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.9.2
altair==4.2.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.12.13
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
#-e git+ssh://git@github.com/vvdiaz1/automatic_tagging.git@b0e6bcbc2f8aaac0f7d583d19a7c138f43fb3946#egg=automatic_tagging
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==22.12.0
bleach==5.0.1
blinker==1.5
botocore==1.27.59
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
coverage==7.0.5
cryptography==39.0.0
cycler==0.11.0
databricks-cli==0.17.4
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker==6.0.1
docutils==0.19
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
fastapi==0.89.1
fastjsonschema==2.16.2
flake8==6.0.0
Flask==2.2.2
flatbuffers==23.1.4
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
gcsfs==2022.11.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio==2.23.0
importlib-metadata==5.2.0
importlib-resources==5.10.1
iniconfig==1.1.1
ipdb==0.13.11
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.11.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.2
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
keras==2.11.0
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
libclang==14.0.6
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
memoized-property==1.0.3
mistune==2.0.4
mlflow==2.1.1
more-itertools==9.0.0
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
pyflakes==3.0.1
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.9
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python-multipart==0.0.5
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
readme-renderer==37.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.1.0
rsa==4.9
s3fs==2022.11.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.3
semver==2.13.0
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.22.0
streamlit==1.17.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
twine==4.0.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.13
uvicorn==0.20.0
validators==0.20.0
watchdog==2.2.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
yapf==0.32.0
yarl==1.8.2
zipp==3.11.0
